refactor: rename ProposalStatusIcon to ProposalIconStatus
