+++ apps/ui/src/components/ProposalsListItem.vue
@@ -34,7 +34,7 @@ async function handleVoteClick(choice: Choice) {
           }"
           class="space-x-2 flex"
         >
-          <ProposalStatusIcon width="17" height="17" :state="proposal.state" class="top-[7.5px]" />
+          <ProposalIconStatus width="17" height="17" :state="proposal.state" class="top-[7.5px]" />
           <h3
             class="leading-6 my-1 text-[21px] md:truncate md:text-ellipsis"
             v-text="proposal.title || `Proposal #${proposal.proposal_id}`"
